+++ jeap-spring-boot-logging-starter-it/src/test/java/ch/admin/bit/jeap/log/syslog/LoggingCloudwatchIT.java
@@ -12,7 +12,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @Slf4j
 @TestPropertySource(properties = "jeap.logging.platform=cloudwatch")
-public class LoggingCloudwatchIT extends SyslogIntegrationTestBase {
+public class LoggingCloudwatchIT extends LogIntegrationTestBase {
 
     @Test
     @StdIo