+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [19.0.0] - 2025-12-03
+
+### Changed
+
+-  Breaking Change
+    - **Removed**
+      - jeap-spring-boot-cloud-autoconfig-starter
+      - jeap-spring-boot-config-starter
+      - other cloudfoundry specifics
+
 ## [18.5.0] - 2025-11-28
 
 ### Changed